Move parser.{h,hpp} out of include/deepstream
