+++ src/error_handler.cpp
@@ -23,7 +23,7 @@
 #include <deepstream/client.hpp>
 #include <deepstream/error_handler.hpp>
 #include <deepstream/message.hpp>
-#include <deepstream/parser.hpp>
+#include "parser.hpp"
 #include "websockets.hpp"
 
 namespace deepstream {
